#4: Removed RLS references in code.
