+++ .coverage_config.lua
@@ -1,6 +1,4 @@
 include = {
-    "main/lua/exasolrls$",
-    "main/lua/exasolrls%/.+$",
     "main/lua/exasolvs$",
     "main/lua/exasolvs%/.+$",
     "main/lua/text$"
