+++ .github/workflows/release_droid_prepare_original_checksum.yml
@@ -1,0 +1,31 @@
+name: Release Droid - Prepare Original Checksum
+
+on:
+  workflow_dispatch:
+
+jobs:
+  build:
+    runs-on: ubuntu-latest
+    steps:
+      - name: Checkout the repository
+        uses: actions/checkout@v3
+        with:
+          fetch-depth: 0
+      - name: Set up JDK 11
+        uses: actions/setup-java@v3
+        with:
+          distribution: 'temurin'
+          java-version: 11
+          cache: 'maven'
+      - name: Enable testcontainer reuse
+        run: echo 'testcontainers.reuse.enable=true' > "$HOME/.testcontainers.properties"
+      - name: Run tests and build with Maven
+        run: mvn --batch-mode clean verify --file pom.xml
+      - name: Prepare checksum
+        run: find target -maxdepth 1 -name *.jar -exec sha256sum "{}" + > original_checksum
+      - name: Upload checksum to the artifactory
+        uses: actions/upload-artifact@v3
+        with:
+          name: original_checksum
+          retention-days: 5
+          path: original_checksum
