+++ .github/workflows/release_droid_print_quick_checksum.yml
@@ -1,0 +1,24 @@
+name: Release Droid - Print Quick Checksum
+
+on:
+  workflow_dispatch:
+
+jobs:
+  build:
+    runs-on: ubuntu-latest
+    steps:
+      - name: Checkout the repository
+        uses: actions/checkout@v3
+        with:
+          fetch-depth: 0
+      - name: Set up JDK 11
+        uses: actions/setup-java@v3
+        with:
+          distribution: 'temurin'
+          java-version: 11
+          cache: 'maven'
+      - name: Build with Maven skipping tests
+        run: mvn --batch-mode clean verify -DskipTests
+      - name: Print checksum
+        run: echo 'checksum_start==';find target -maxdepth 1 -name *.jar -exec sha256sum "{}" + | xargs;echo '==checksum_end'
+
