--- doc/model/diagrams/component/comp_rls_administration_scripts.plantuml
@@ -1,38 +1,0 @@
-@startuml
-!include ../clean.skin
-
-title: Collection of RLS administration scripts
-
-package "Administration Script Collection" as script {
-    ' [dsn -> req~assigning-users-to-groups~1 >> impl, itest]
-    component add_user_to_group
-
-    ' [dsn -> req~removing-users-from-groups~1 >> impl, itest]
-    component remove_user_from_group
-
-    ' [dsn -> req~listing-all-groups~1 >> impl, itest]
-    component list_all_groups
-    
-    ' [dsn -> req~listing-a-users-groups~1 >> impl, itest]
-    component list_user_groups
-
-    ' [dsn -> req~assigning-roles-to-users~1 >> impl, itest]
-    component assign_roles_to_user
-    component add_rls_role
-
-    ' [dsn -> req~removing-roles-from-users~1 >> impl, itest]
-    component delete_rls_role
-
-    component bit_positions 
-
-    ' [dsn -> req~listing-all-roles~1 >> impl, itest]
-    component list_all_roles
-
-    ' [dsn -> req~listing-users-and-roles~1 >> impl, itest]
-    component list_users_and_roles
-    
-    ' [dsn -> req~listing-user-roles~1 >> impl, itest]
-    component list_user_roles
-}
-
-@enduml
