+++ doc/model/diagrams/sequence/seq_pushdown.plantuml
@@ -1,37 +1,36 @@
+' [dsn -> req~push-down~1]
+
 @startuml
 !include ../clean.skin
 
 title: Query Push-down
 
-Participant request_dispatcher
-Participant rls_adapter
-Participant query_rewriter
-Participant table_protection_reader
-Participant query_renderer
+Participant RequestDispatcher
+Participant ExasolAdapter
+Participant QueryRewriter
+Participant QueryRenderer
 
-activate request_dispatcher
-request_dispatcher -> rls_adapter : push down (request)
-activate rls_adapter
-rls_adapter -> rls_adapter : extract adapter notes (request)
-activate rls_adapter
-deactivate rls_adapter
-rls_adapter -> query_rewriter : rewrite (query, schema\nadapter notes, involved table[])
-activate query_rewriter
-query_rewriter -> table_protection_reader : read (adapter notes)
-activate table_protection_reader 
-table_protection_reader -->> query_rewriter : table protection scheme[] 
-deactivate table_protection_reader
+activate RequestDispatcher
+RequestDispatcher -> ExasolAdapter : push down (request)
+activate ExasolAdapter
+ExasolAdapter -> ExasolAdapter : extract adapter notes (request)
+activate ExasolAdapter
+deactivate ExasolAdapter
+ExasolAdapter -> QueryRewriter : rewrite (query, schema\nadapter notes, involved table[])
+activate QueryRewriter
+ref over QueryRewriter : **[[../activity/act_rewrite_with_protection.svg rewrite query]]**
 
-ref over query_rewriter : **[[../activity/act_rewrite_with_protection.svg rewrite query]]** 
-
-create query_renderer
-query_rewriter -> query_renderer : create (query)
-query_rewriter -> query_renderer : render ()
-activate query_renderer
-query_renderer -->> query_rewriter : query : SQL string
-deactivate query_renderer
-query_rewriter --> rls_adapter : query : SQL string
-deactivate query_rewriter
-rls_adapter -->> request_dispatcher : wrapped response
-deactivate rls_adapter 
+create QueryRenderer
+QueryRewriter -> QueryRenderer : new(query)
+activate QueryRenderer
+QueryRenderer -->> QueryRewriter : instance
+deactivate QueryRenderer
+QueryRewriter -> QueryRenderer : render ()
+activate QueryRenderer
+QueryRenderer -->> QueryRewriter : query : SQL string
+deactivate QueryRenderer
+QueryRewriter --> ExasolAdapter : query : SQL string
+deactivate QueryRewriter
+ExasolAdapter -->> RequestDispatcher : wrapped response
+deactivate ExasolAdapter
 @enduml
