+++ doc/model/diagrams/sequence/seq_refresh.plantuml
@@ -1,29 +1,22 @@
-' [dsn->req~refreshing-an-rls-virtual-schema~1]
+' [dsn->req~refreshing-a-virtual-schema~1]
 ' [dsn->req~filtering-tables~1]
 
 @startuml
 !include ../clean.skin
 
-title: Refresch Virtual Schema
+title: Refresh Virtual Schema
 
-Participant request_dispatcher
-Participant rls_adapter
+Participant RequestDispatcher
+Participant ExasolAdapter
 
-activate request_dispatcher
-request_dispatcher -> rls_adapter : refresh (request, table[])
-activate rls_adapter
-rls_adapter -> metadata_reader : read (schema)
+activate RequestDispatcher
+RequestDispatcher -> ExasolAdapter : refresh (request, table[])
+activate ExasolAdapter
+ExasolAdapter -> metadata_reader : read (schema)
 activate metadata_reader
 ref over metadata_reader : **[[seq_read_table_metadata.svg read table metadata]]**
-metadata_reader -->> rls_adapter : schema metadata,\ntable protection type
+metadata_reader -->> ExasolAdapter : schema metadata
 deactivate metadata_reader
-rls_adapter -->> request_dispatcher : schema metadata, adapter notes
-note right
-    Response contains information
-    about protection scheme for each
-    table as adapter notes.
-    Adapter notes serve as cache
-    across requests.
-end note
-deactivate rls_adapter
+ExasolAdapter -->> RequestDispatcher : schema metadata, adapter notes
+deactivate ExasolAdapter
 @enduml
