+++ src/main/lua/entry.lua
@@ -1,9 +1,9 @@
 --- Main entry point of the Lua Virtual Schema adapter.
 -- It is responsible for creating and wiring up the main adapter objects.
 
-local RlsAdapter = require("exasolrls.RlsAdapter")
-local RlsAdapterProperties = require("exasolrls.RlsAdapterProperties")
-local MetadataReader = require("exasolrls.MetadataReader")
+local ExasolAdapter = require("exasolvs.ExasolAdapter")
+local ExasolAdapterProperties = require("exasolvs.ExasolAdapterProperties")
+local MetadataReader = require("exasolvs.MetadataReader")
 local RequestDispatcher = require("exasolvs.RequestDispatcher")
 
 --- Handle a Virtual Schema request.
@@ -12,7 +12,7 @@ local RequestDispatcher = require("exasolvs.RequestDispatcher")
 function adapter_call(request_as_json)
     local exasol_context = _G.exa
     local metadata_reader = MetadataReader:new(exasol_context)
-    local adapter = RlsAdapter:new(metadata_reader)
-    local dispatcher = RequestDispatcher:new(adapter, RlsAdapterProperties)
+    local adapter = ExasolAdapter:new(metadata_reader)
+    local dispatcher = RequestDispatcher:new(adapter, ExasolAdapterProperties)
     return dispatcher:adapter_call(request_as_json)
 end
