+++ src/main/lua/exasolvs/ExasolAdapter.lua
@@ -1,63 +1,62 @@
 local AbstractVirtualSchemaAdapter = require("exasolvs.AbstractVirtualSchemaAdapter")
-local adapter_capabilities = require("exasolrls.adapter_capabilities")
-local QueryRewriter = require("exasolrls.QueryRewriter")
+local adapter_capabilities = require("exasolvs.adapter_capabilities")
+local QueryRewriter = require("exasolvs.QueryRewriter")
 
--- Derive from AbstractVirtualSchemaAdapter
-local RlsAdapter = {}
-RlsAdapter.__index = RlsAdapter
-setmetatable(RlsAdapter, {__index = AbstractVirtualSchemaAdapter})
-local VERSION <const> = "1.2.0"
+--- Virtual Schema adapter for Exasol-to-Exasol Virtual Schemas.
+-- @classmod ExasolAdapter
+local ExasolAdapter = {}
+ExasolAdapter.__index = ExasolAdapter
+setmetatable(ExasolAdapter, {__index = AbstractVirtualSchemaAdapter})
+local VERSION <const> = "0.1.0"
 
---- Create an `RlsAdapter`.
+--- Create an `ExasolAdapter`.
 -- @param metadata_reader metadata reader
--- @return RlsAdapter
-function RlsAdapter:new(metadata_reader)
+-- @return new instance
+function ExasolAdapter:new(metadata_reader)
     local instance = setmetatable({}, self)
     instance:_init(metadata_reader)
     return instance
 end
 
-function RlsAdapter:_init(metadata_reader)
+function ExasolAdapter:_init(metadata_reader)
     AbstractVirtualSchemaAdapter._init(self)
     self._metadata_reader = metadata_reader
 end
 
 --- Get the version number of the Virtual Schema adapter.
 -- @return Virtual Schema adapter version
-function RlsAdapter:get_version()
+function ExasolAdapter:get_version()
     return VERSION
 end
 
 --- Get the name of the Virtual Schema adapter.
 -- @return Virtual Schema adapter name
-function RlsAdapter:get_name()
-    return "Row-level Security adapter (Lua)"
+function ExasolAdapter:get_name()
+    return "Exasol Virtual Schema adapter (Lua)"
 end
 
 --- Create a virtual schema.
 -- @param request virtual schema request
 -- @param properties user-defined properties
 -- @return response containing the metadata for the virtual schema like table and column structure
-function RlsAdapter:create_virtual_schema(request, properties)
+function ExasolAdapter:create_virtual_schema(request, properties)
     properties:validate()
     local metadata = self:_handle_schema_scanning_request(request, properties)
     return {type = "createVirtualSchema", schemaMetadata = metadata}
 end
 
-function RlsAdapter:_handle_schema_scanning_request(request, properties)
+function ExasolAdapter:_handle_schema_scanning_request(_, properties)
     local schema_name = properties:get_schema_name()
     local table_filter = properties:get_table_filter()
     return self._metadata_reader:read(schema_name, table_filter)
 end
 
 --- Refresh the metadata of the Virtual Schema.
--- <p>
 -- Re-reads the structure and data types of the schema protected by RLS.
--- </p>
 -- @param request virtual schema request
 -- @param properties user-defined properties
 -- @return response containing the metadata for the virtual schema like table and column structure
-function RlsAdapter:refresh(request, properties)
+function ExasolAdapter:refresh(request, properties)
     properties:validate()
     return {type = "refresh", schemaMetadata = self:_handle_schema_scanning_request(request, properties)}
 end
@@ -66,7 +65,7 @@ end
 -- @param request virtual schema request
 -- @param properties user-defined properties
 -- @return response containing the metadata for the virtual schema like table and column structure
-function RlsAdapter:set_properties(request, properties)
+function ExasolAdapter:set_properties(request, properties)
     properties:validate()
     return {type = "setProperties", schemaMetadata = self:_handle_schema_scanning_request(request, properties)}
 end
@@ -75,7 +74,7 @@ end
 -- @param request virtual schema request
 -- @param properties user-defined properties
 -- @return response containing the list of reported capabilities
-function RlsAdapter:push_down(request, properties)
+function ExasolAdapter:push_down(request, properties)
     properties:validate()
     local adapter_cache = request.schemaMetadataInfo.adapterNotes
     local rewritten_query = QueryRewriter.rewrite(request.pushdownRequest, properties:get_schema_name(),
@@ -83,8 +82,8 @@ function RlsAdapter:push_down(request, properties)
     return {type = "pushdown", sql = rewritten_query}
 end
 
-function RlsAdapter:_define_capabilities()
+function ExasolAdapter:_define_capabilities()
     return adapter_capabilities
 end
 
-return RlsAdapter
+return ExasolAdapter
