+++ src/main/lua/exasolvs/ExasolAdapterProperties.lua
@@ -3,21 +3,21 @@ local exaerror = require("exaerror")
 local AdapterProperties = require("exasolvs.AdapterProperties")
 
 --- This class abstracts access to the user-defined properties of the Virtual Schema.
--- @classmod RlsAdapterProperties
-local RlsAdapterProperties = {}
-RlsAdapterProperties.__index = RlsAdapterProperties
-setmetatable(RlsAdapterProperties, AdapterProperties)
+-- @classmod ExasolAdapterProperties
+local ExasolAdapterProperties = {}
+ExasolAdapterProperties.__index = ExasolAdapterProperties
+setmetatable(ExasolAdapterProperties, AdapterProperties)
 
---- Create a new `RlsAdapterProperties` instance
+--- Create a new `ExasolAdapterProperties` instance
 -- @param raw_properties unparsed user-defined properties
--- @return new RLS adaper properties
-function RlsAdapterProperties:new(raw_properties)
+-- @return new instance
+function ExasolAdapterProperties:new(raw_properties)
     local instance = setmetatable({}, self)
     instance:_init(raw_properties)
     return instance
 end
 
-function RlsAdapterProperties:_init(raw_properties)
+function ExasolAdapterProperties:_init(raw_properties)
     AdapterProperties._init(self, raw_properties)
 end
 
@@ -26,7 +26,7 @@ local TABLE_FILTER_PROPERTY <const> = "TABLE_FILTER"
 
 --- Validate the adapter properties.
 -- @raise validation error
-function RlsAdapterProperties:validate()
+function ExasolAdapterProperties:validate()
     AdapterProperties.validate(self) -- super call
     if not self:has_value(SCHEMA_NAME_PROPERTY) then
         exaerror.create("F-RLS-PROP-1", "Missing mandatory property '" .. SCHEMA_NAME_PROPERTY .. "' ")
@@ -41,15 +41,15 @@ end
 
 --- Get the name of the Virtual Schema's source schema.
 -- @return name of the source schema
-function RlsAdapterProperties:get_schema_name()
+function ExasolAdapterProperties:get_schema_name()
     return self:get(SCHEMA_NAME_PROPERTY)
 end
 
 --- Get the list of tables that the Virtual Schema should show after applying the table filter.
 -- @return list of tables
-function RlsAdapterProperties:get_table_filter()
+function ExasolAdapterProperties:get_table_filter()
     local filtered_tables = self:get(TABLE_FILTER_PROPERTY)
     return text.split(filtered_tables)
 end
 
-return RlsAdapterProperties
+return ExasolAdapterProperties
