+++ src/main/lua/exasolvs/adapter_capabilities.lua
