+++ src/test/java/com/exasol/ExasolVirtualSchemaTestConstants.java
@@ -1,0 +1,9 @@
+package com.exasol;
+
+public final class ExasolVirtualSchemaTestConstants {
+    public static final String DOCKER_DB = "exasol/docker-db:7.1.10";
+
+    private ExasolVirtualSchemaTestConstants() {
+        // prevent instantiation
+    }
+}
